extract `combine` functions
